[fix] : 일정 서비스 코드 구현 / 선언으로 나누기
Resolves : #26
