+++ src/main/java/backend/zip/service/ScheduleService.java
@@ -1,47 +1,17 @@
 package backend.zip.service;
 
 import backend.zip.domain.schedule.Schedule;
-import backend.zip.domain.user.User;
-import backend.zip.repository.ScheduleRepository;
-import backend.zip.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
+import backend.zip.dto.schedule.request.AddScheduleRequest;
+import backend.zip.dto.schedule.request.UpdateScheduleRequest;
 
 import java.util.Optional;
 
-@Service
-@RequiredArgsConstructor
-public class ScheduleService {
+public interface ScheduleService {
+    Optional<Schedule> getScheduleByUserId(Long userId);
 
-    private final ScheduleRepository scheduleRepository;
-    private final UserRepository userRepository;
+    Schedule addSchedule(Long userId, AddScheduleRequest request);
 
-    //조회
-    public Optional<Schedule> getScheduleByUserId(Long userId) {
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new IllegalArgumentException("사용자를 찾을 수 없습니다: "));
+    Optional<Schedule> updateSchedule(Long userId, UpdateScheduleRequest request);
 
-        return scheduleRepository.findByUser(user);
-    }
-
-    //생성
-    public void createSchedule(Schedule schedule) {
-        scheduleRepository.save(schedule);
-    }
-
-    //수정
-    public Optional<Schedule> updateSchedule(Long userId, Schedule updatedSchedule) {
-        return getScheduleByUserId(userId).map(existingSchedule -> {
-            existingSchedule.setMoveDate(updatedSchedule.getMoveDate());
-            existingSchedule.setPeriod(updatedSchedule.getPeriod());
-            return scheduleRepository.save(existingSchedule);
-        });
-    }
-
-    //삭제
-    public void deleteSchedule(Long userId) {
-        User deleteUser = userRepository.findById(userId).get();
-        scheduleRepository.deleteByUser(deleteUser);
-    }
-
+    void deleteSchedule(Long userId);
 }
